+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - For templates which use Paged.js, lines can be numbered using the top-level YAML parameter `number-lines` (thanks, @julientaq, #115 and #129).
 
+## BUG FIXES
+
+- Fixed several bugs related to MathJax: local version of MathJax is now used when the `mathjax` parameter is set to `"local"` and self contained documents are rendered by default with MathJax without throwing any warning (#130).
+
 # CHANGES IN pagedown VERSION 0.3
 
 ## NEW FEATURES